Started Debugging and also changed the folder names to the Functional Requirements name
